+++ packages/react-native-attestation/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @bifold/react-native-attestation
 
+## 2.7.2
+
+### Patch Changes
+
+- 99dab08: added missing react imports
+
+## 2.7.1
+
 ## 2.7.0
 
 ## 2.6.0